src/requirements.txt
flask-restful==0.3.5
Flask-SQLAlchemy==2.1
flask-MySQLdb==0.2.0
flask-cors==2.1.2
marshmallow==2.4.2
requests==2.9.1
py-bcrypt==0.4
PyYAML==3.11
gunicorn==19.4.5
python-zimbra==2.0
celery==3.1.19
redis==2.10.5
PyJWT==1.4.0
